ボスのHP表示（全体で一個を共用）、playerのhp位置変更、警告を改善。

+++ EnemyManager.cpp
@@ -151,7 +151,7 @@ void EnemyManager::UpdateEnemyPopCommands()
 
 			//hp
 			getline(line_stream, word, ',');
-			int HP = (float)std::atof(word.c_str());
+			int HP = (int)std::atof(word.c_str());
 
 			BossGenerate({ x,y,z }, scale, HP);
 		}
